Build 385 (working in Debug; maybe buildable): dev of the texture assets browser #2; refactoring of some editor elements

+++ Core/Render/graphicsclass.cpp
@@ -1013,7 +1013,7 @@ void GraphicsClass::SetupLightsForFrame(
     const size numVisPointLightSources           = visPointLights.size();
 
 
-    outData.ResizeLightData(numDirLights, (int)numVisPointLightSources, numSpotLights);
+    outData.ResizeLightData((int)numDirLights, (int)numVisPointLightSources, (int)numSpotLights);
 
     if (numVisPointLightSources > 0)
     {
